refactor generator, do pst ast full roundtrip
