+++ cedar-drt/fuzz/fuzz_targets/pst-est-roundtrip.rs
@@ -17,51 +17,19 @@
 #![no_main]
 
 use cedar_drt::logger::initialize_log;
-use cedar_drt_inner::{fuzz_target, pst_equiv};
+use cedar_drt_inner::{fuzz_target, pst_equiv, pst_gen::FuzzTargetInput};
 
 use cedar_policy_core::est;
 use cedar_policy_core::pst;
-use cedar_policy_generators::hierarchy::HierarchyGenerator;
-use cedar_policy_generators::pst::{arbitrary_pst_template, arbitrary_pst_template_size_hint};
-use cedar_policy_generators::schema::Schema;
-use cedar_policy_generators::schema_gen::SchemaGen;
-use cedar_policy_generators::settings::ABACSettings;
-use libfuzzer_sys::arbitrary::{self, Arbitrary, Unstructured};
 use log::debug;
 
-/// Fuzz target input: a PST template generated from a schema/hierarchy.
-#[derive(Debug, Clone)]
-struct FuzzTargetInput {
-    template: pst::Template,
-}
-
-const SETTINGS: ABACSettings = ABACSettings {
-    enable_additional_attributes: true,
-    enable_arbitrary_func_call: false,
-    ..ABACSettings::undirected()
-};
-
-impl<'a> Arbitrary<'a> for FuzzTargetInput {
-    fn arbitrary(u: &mut Unstructured<'a>) -> arbitrary::Result<Self> {
-        let schema = Schema::arbitrary(SETTINGS.clone(), u)?;
-        let hierarchy = schema.arbitrary_hierarchy(u)?;
-        let template = arbitrary_pst_template(&hierarchy, 3, 3, u)?;
-        Ok(Self { template })
-    }
-
-    fn try_size_hint(
-        depth: usize,
-    ) -> arbitrary::Result<(usize, Option<usize>), arbitrary::MaxRecursionReached> {
-        Ok(arbitrary::size_hint::and_all(&[
-            Schema::arbitrary_size_hint(depth)?,
-            HierarchyGenerator::size_hint(depth),
-            arbitrary_pst_template_size_hint(depth),
-        ]))
-    }
-}
-
 // PST → EST → PST roundtrip
-fn round_trip_pst_est(template: &pst::Template) -> pst::Template {
+fuzz_target!(|input: FuzzTargetInput| {
+    initialize_log();
+    let template = input.template;
+
+    debug!("Running PST→EST→PST roundtrip on: {:?}", template);
+
     let est_policy: est::Policy = template
         .clone()
         .try_into()
@@ -71,17 +39,7 @@ fn round_trip_pst_est(template: &pst::Template) -> pst::Template {
         .try_into()
         .unwrap_or_else(|e| panic!("EST → PST failed: {:?}\nOriginal: {:?}", e, template));
 
-    roundtripped
-}
-
-fuzz_target!(|input: FuzzTargetInput| {
-    initialize_log();
-    let template = input.template;
-
-    debug!("Running PST→EST→PST roundtrip on: {:?}", template);
-
-    let roundtripped = round_trip_pst_est(&template);
-    // Check PST equivalence, not including IDs because they get lost in the EST
+    // IDs are lost in the EST roundtrip
     pst_equiv::check_template_equivalence(
         &template,
         &roundtripped,
